From the example above, we are able to tell what column contains which information. This is made possible with the use of <th> tag
How to Add a Table Heading in HTML
The <th> is used to add headings to tables. In basic designs the table heading will always take the top row, meaning we will have the <th> declared in our first table row followed by the actual data in the table. By default the text passed in the Heading is centered and Bold.